# Conflicts:
#	data_storage/for_backup_storage/2/forecastJJOralCare/default.pickle

+++ iap/common/repository/models_managers/scenario_manager.py
@@ -1,6 +1,7 @@
 from ..models.scenarios import Scenario
 from ..models.access import User
 from sqlalchemy.orm.exc import NoResultFound
+from ..models_managers import access_manager
 import transaction
 import datetime
 
@@ -19,7 +20,8 @@ def create_scenario(session, input_data, user=None):
         date_of_last_mod = str(datetime.datetime.now())
         scenario = Scenario(name=input_data['name'], description=input_data['description'],
                             shared=input_data['shared'], date_of_last_modification=date_of_last_mod,
-                            status="New", criteria=input_data['criteria'], author=input_data['author'])
+                            status=input_data['status'], criteria=input_data['criteria'], author=input_data['author'],
+                            )
         session.add(scenario)
         if user:
             user.scenarios.append(scenario)
@@ -29,16 +31,13 @@ def create_scenario(session, input_data, user=None):
     return scenario
 
 
-def get_scenario_by_id(session,user_id, scenario_id):
+def get_scenario_by_id(session, user_id, scenario_id):
 
     query = session.query(Scenario)
     query = query.join(User.scenarios)
     query = query.filter(Scenario.id == scenario_id)
     query = query.filter(User.id == user_id)
     scenario = query.one()
-    #scenario = query.filter(User.id == user_id).one()
-    #else:
-        # TODO join with user_id
     return scenario
 
 
@@ -104,6 +103,15 @@ def include_scenario(session, user_id, scenario_id, parent_scenario_id):
     return
 
 
+def update_scenario(scenario, parmeter, value):
+
+    if getattr(scenario, parmeter) == value:
+        pass
+    else:
+        setattr(scenario, parmeter, value)
+    return
+
+
 def delete_scenario(session, scenario_id, user_id):
     """
     Delete scenario's
@@ -117,7 +125,14 @@ def delete_scenario(session, scenario_id, user_id):
     :rtype:
     """
     try:
-        scenario = get_scenario_by_id(session, scenario_id, user_id)
-        session.delete(scenario)
+        user = access_manager.get_user_by_id(session, user_id=user_id)
+        scenario = get_scenario_by_id(session, scenario_id=scenario_id, user_id=user_id)
+        if scenario.author == user.email:
+            session.delete(scenario)
+            status = "Deleted"
+        else:
+            status = "Unauthorised"
     except NoResultFound:
-        raise Exception
+        return "No item"
+    else:
+        return status
